networth_tracker/apis.py
from rest_framework import permissions, response, views

from .api import serializers as user_serializer


class RegisterView(views.APIView):
    permission_classes = (permissions.AllowAny,)

    def post(self, request):
        serializer = user_serializer.CustomUserSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        serializer.save()

        return response.Response(serializer.data)
