requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
whitenoise==6.6.0
